# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for GroceryTrak.",
        "title": "GroceryTrak API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GroceryTrak",
            "email": "grocerytrak@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/user_item": {
            "get": {
                "description": "Get all items for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_item"
                ],
                "summary": "Get all user's items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemsResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new item for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_item"
                ],
                "summary": "Create a new user_item for the authenticated user",
                "parameters": [
                    {
                        "description": "Create User Item",
                        "name": "userItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user_item/predict": {
            "post": {
                "description": "Predict items from an uploaded image for the authenticated user",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_item"
                ],
                "summary": "Predict items from an uploaded image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemsResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user_item/search": {
            "get": {
                "description": "Searches for user items by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_item"
                ],
                "summary": "Search user items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of user item",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemsResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user_item/{item_id}": {
            "get": {
                "description": "Get a specific item for the authenticated user by item ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_item"
                ],
                "summary": "Get a user's item by ItemID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a user_item for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_item"
                ],
                "summary": "Update a user_item for the authenticated user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update User Item",
                        "name": "userItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.UserItemResponse"
                        }
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a user_item for the authenticated user",
                "tags": [
                    "user_item"
                ],
                "summary": "Delete a user_item for the authenticated user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "default": {
                        "description": "Standard Error Responses",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
...
        "dtos.ItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "milk.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "Milk"
                },
                "spoonacular_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "dtos.UserItemRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 2
                },
                "item_id": {
                    "type": "integer",
                    "example": 456
                },
                "unit": {
                    "type": "string",
                    "example": "kg"
                }
            }
        },
        "dtos.UserItemResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 2
                },
                "item": {
                    "$ref": "#/definitions/dtos.ItemResponse"
                },
                "unit": {
                    "type": "string",
                    "example": "kg"
                }
            }
        },
        "dtos.UserItemsResponse": {
            "type": "object",
            "properties": {
                "user_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.UserItemResponse"
                    }
                }
            }
        }
    }
}